+++ src/config/config.ts
@@ -1,8 +1,10 @@
+import { Comparable } from "@/comparision/comparable";
 import { ConfigLoader } from "@/config/configLoader";
 import { PropagationTestConfig } from "@/config/propagationTestConfig";
 import { ScenarioCompileConfig } from "@/config/scenarioCompileConfig";
 import { IFile } from "@/fs/IFile";
 import { FsContainer } from "@/fs/fsContainer";
+import { Method } from "@/type/http";
 
 export const ECHOED_CONFIG_FILE_NAME = ".echoed.yml";
 
@@ -32,9 +34,37 @@ export type ServiceConfig = {
 export type OpenApiConfig = {
   filePath: string;
   basePath?: string;
+  coverage?: OpenApiCoverageConfig;
+};
+
+export type OpenApiCoverageConfig = {
+  undocumentedOperation: OpenApiUndocumentedOperationConfig;
+};
+
+export type OpenApiUndocumentedOperationConfig = {
+  ignores: OpenApiIgnoreOperationConfig[];
+};
+
+export type OpenApiIgnoreOperationConfig = {
+  path: Comparable;
+  method: Method;
 };
 
 export type ProtoConfig = {
   filePath: string;
   services?: string[];
+  coverage?: ProtoCoverageConfig;
+};
+
+export type ProtoCoverageConfig = {
+  undocumentedMethod: ProtoUndocumentedMethodConfig;
+};
+
+export type ProtoUndocumentedMethodConfig = {
+  ignores: ProtoIgnoreMethodConfig[];
+};
+
+export type ProtoIgnoreMethodConfig = {
+  service: Comparable;
+  method: Comparable;
 };